// apps/desktop/src/types/generated/WallpaperCollection.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { WallpaperCatalogEntry } from "./WallpaperCatalogEntry";
import type { WallpaperScanStats } from "./WallpaperScanStats";

export type WallpaperCollection = { entries: Array<WallpaperCatalogEntry>, generated_at: string, scan_stats: WallpaperScanStats, };