// apps/desktop/src/types/generated/WallpaperFilter.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { WallpaperKind } from "./WallpaperKind";
import type { WallpaperSource } from "./WallpaperSource";

export type WallpaperFilter = { kind: WallpaperKind | null, source: WallpaperSource | null, };