// apps/desktop/src/types/generated/CurrentWallpaperState.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { WallpaperConfidence } from "./WallpaperConfidence";
import type { WallpaperId } from "./WallpaperId";

export type CurrentWallpaperState = { last_applied_by_app: WallpaperId | null, reported_by_backend: string | null, checked_at: string, confidence: WallpaperConfidence, };